pyproject.toml
[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
line_length = 100
profile = 'black'

[tool.coverage.run]
omit = ["*/test_*.py"]

[tool.poetry]
name = "sec_ai"
version = "1.0.0"
description = "Security AI"
authors = ["AI Team <ai-team@allisone.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.5"
pandas = "1.4.2"
boto3 = "1.22.10"
opencv-python = "4.5.5.64"
pillow = "9.1.0"
darwin-py = "0.8.6"
shapely = "1.8.0"
tensorboard = "2.7.0"
scikit-learn = "1.0.2"
plotly = "5.7.0"
albumentations = "1.1.0"
setuptools = "59.5.0"
protobuf = "3.20.0"
rasterio = "^1.3.4"
matplotlib = "^3.6.2"
python-dotenv = "^0.21.0"
pycocotools = "^2.0.6"
ipdb = "^0.13.11"
pytest = "^7.2.0"
ipython = "^8.8.0"
torch-tb-profiler = "^0.4.1"
omegaconf = "^2.3.0"
tqdm = "^4.64.1"
torchmetrics = "^0.11.0"
torch = { url = "https://download.pytorch.org/whl/cu116/torch-1.13.1%2Bcu116-cp310-cp310-linux_x86_64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/cu116/torchvision-0.14.1%2Bcu116-cp310-cp310-linux_x86_64.whl" }
transformers = "4.30.1"
pytorchvideo = "^0.1.5"

[tool.poetry.group.dev.dependencies]
clearml = "*"
pylint = "^2.15.10"
black = "23.7.0"
ipdb = "*"
pytest = "*"
pytest-cov = "*"
pre-commit = "2.20.0"
ipykernel = "^6.20.1"
jupyter = "^1.0.0"
flake8 = "^6.0.0"
